requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
paho-mqtt==1.6.1
mysql-connector-python==8.2.0
python-dotenv==1.0.0
pydantic==2.5.3
numpy
pandas
scikit-learn
joblib